Bug in PCB: Changed One wire DS18b20 sensor from A7 to A4
There is an error in the PCB, the one wire temperature sensor DS18b20 cannot work on analog port A7 (the A6 and A7 IO pins can only be used as analog pins, not as digital IO). So you have to use A4, or A5 to connect the one wire sensor.